+++ pycondor/tests/test_dagman.py
@@ -179,7 +179,8 @@ def test_dagman_job_order(tmpdir):
 def test_repr():
     default_dagman = Dagman('dagname')
     dag_repr = repr(default_dagman)
-    expected_repr = 'Dagman(name=dagname, n_nodes=0)'
+    expected_repr = ('Dagman(name=dagname, n_nodes=0, '
+                     'submit={})'.format(os.getcwd()))
     assert dag_repr == expected_repr
 
     dag_non_default = Dagman('dagname', submit='/submit_dir')
@@ -228,7 +229,7 @@ def test_dagman_env_variable_dir(tmpdir, monkeypatch):
 
     # Set pycondor environment variable
     submit_dir = str(tmpdir.mkdir('submit'))
-    monkeypatch.setattr(os, 'getenv', lambda *args: submit_dir)
+    monkeypatch.setenv('PYCONDOR_SUBMIT_DIR', submit_dir)
 
     dagman = Dagman('testdagman')
     job = Job('jobname', example_script)